requirements.txt
django>=1.4,<1.5
pip
numpy
scipy
Markdown<2.3.0
pygments
lxml
boto
mako
python-memcached
python-openid 
path.py
django_debug_toolbar
fs
beautifulsoup 
beautifulsoup4
feedparser
requests
sympy
newrelic
glob2
pymongo
django_nose
nosexcover
rednose
GitPython >= 0.3
django-override-settings
mock>=0.8, <0.9
PyYAML
South
pytz
django-celery
django-countries
django-kombu
django-followit
django-jasmine
django-keyedcache
django-mako
django-masquerade
django-openid-auth 
django-robots
django-ses
django-storages
django-threaded-multihost
django-sekizai<0.7
django-mptt>=0.5.3
sorl-thumbnail
networkx
pygraphviz
-r repo-requirements.txt
pil
